Reasonable solution that doesn't seem to work.
I've looked at the visual and can verify the data seems
to be as wanted. It just doesn't seem to learn, I don't
have any good ideas why. The results are always about
10% correct; what I'd expect from a completely random guess.
It might not be the greatest implementation, but I'd be
surprised if there's no bugs and it can't do better than
random.

Also, it seems to get locked in to results sometimes. For example,
right now it's scoring 102/1000 every time. It's the same test
data, but different training data as we go. It doesn't make
sense that the results would be so exact and consistent; weights
should be causing variance.