Added evalution of pixel density to input data.
Wrote method get_densities() to give an estimate of the
pixel density in 5 parts of the image: total, north of horizontal,
south of horizontal, west of vertical, and east of vertical.

Added this to the inputs to see if it helps learning at all.

Still playing with multipliers, not sure what a good magnitude for
these numbers would be. I think it's just important that there's
enough of a spread for the network to differentiate.