Organized into a proper app to run experiment and training.
I'm really excited about this actually, it's gonna be great.
The various functions are differentiated via command line
parameters. The experiment can run in its entirety and generate
a graph immediately following (versus a separate app before).

Training will automatically look for the best known weight
configuration and load it, or start from scratch if it doesn't
exist. Whenever it bests itself, it will save the weights and
record the improvement.

This was probably too much change to properly put in a single
commit, but... it's awesome now so it's OK. I think awesome
anyway, testing was kinda half-assed so far.