Added some output and sped up learning.
I found myself wanting more output watching, so
it prints more status updates to stdout. There's no
more spam in the log though.

I also cut the sample rate for learning down by a lot,
it just took SO LONG and I don't think the benefit
(if any, 3,000 is still a lot of samples) is
worth the time. I'm impatient.