Solid, working application.
The basics are in place. Could stand to be cleaned up a bit.
I think there are memory leaks and some code may be better
refactored.

The accuracy is good when the character is centered and sized
similarly to the data on which the network was trained. To
help improve the performance outside of those cases I'm going
to add perturbations of the training data to the training data
(minor scales and rotations).